refactor: 매직 넘버 상수화 및 util 파일 분리
