+++ src/feature/add-log/components/StepComplete.tsx
@@ -1,4 +1,4 @@
-import { postMessageToWebView } from '@/utils';
+import { postMessageToWebView } from '@/utils/webview';
 
 function StepComplete() {
   const routeHome = () => {
